AUI-1328 Split Gulp tasks into separated files
